poc/requirements.txt
# Shared dependencies for all POCs
# Install: pip install -r requirements.txt

# === Core Framework ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20

# === LLM & Embeddings ===
litellm==1.57.4
tiktoken==0.8.0
openai==1.59.3

# === Vector Database ===
qdrant-client==1.12.1

# === Document Parsing ===
PyMuPDF==1.25.3
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
openpyxl==3.1.5
python-pptx==1.0.2
markdown==3.7

# === NLP & Text Processing ===
nltk==3.9.1
regex==2024.11.6

# === Database ===
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0

# === Task Queue ===
celery[redis]==5.4.0
redis==5.2.1

# === Object Storage ===
boto3==1.36.2
minio==7.2.12

# === Reranking ===
cohere==5.13.4

# === Utilities ===
python-dotenv==1.0.1
xxhash==3.5.0
aiofiles==24.1.0
httpx==0.28.1
json-repair==0.30.3
numpy==1.26.4
uuid6==2024.7.10
rich==13.9.4
